refactor: Separate storage into its own package
