+++ module.go
@@ -11,6 +11,7 @@ import (
 	"github.com/google/tink/go/aead"
 	"github.com/google/tink/go/insecurecleartextkeyset"
 	"github.com/google/tink/go/keyset"
+	certmagicgcs "github.com/grafana/certmagic-gcs/storage"
 )
 
 // Interface guards
@@ -43,7 +44,7 @@ func (CaddyStorageGCS) CaddyModule() caddy.ModuleInfo {
 
 // CertMagicStorage returns a cert-magic storage.
 func (s *CaddyStorageGCS) CertMagicStorage() (certmagic.Storage, error) {
-	config := StorageConfig{
+	config := certmagicgcs.StorageConfig{
 		BucketName: s.BucketName,
 	}
 
@@ -68,7 +69,7 @@ func (s *CaddyStorageGCS) CertMagicStorage() (certmagic.Storage, error) {
 		}
 		config.AEAD = kp
 	}
-	return NewStorage(context.Background(), config)
+	return certmagicgcs.NewStorage(context.Background(), config)
 }
 
 // Validate caddy gcs storage configuration.
